requirements.txt
opencv-python
numpy
Pillow
imageio 